api-service/requirements.txt
azure-functions
openai
pydantic
psycopg2-binary
python-dotenv
langgraph
langsmith